DeepLearningModelDeployment/scoring-pipeline/requirements.txt
filelock==2.0.13
numpy==1.14.5
pandas==0.22.0
pycryptodome==3.6.3
requests==2.18.4
scikit-learn==0.19.1
scipy==1.1.0
setproctitle==1.1.10
statsmodels==0.8.0
toml==0.9.3.1
datatable-0.6.0.dev166-cp36-cp36m-linux_x86_64.whl
h2o4gpu-0.2.0.9999+master.eb6295c-cp36-cp36m-linux_x86_64.whl
h2oaicore-1.2.2-cp36-cp36m-linux_x86_64.whl
scoring_h2oai_experiment_dovudusu-1.0.0-py3-none-any.whl
